+++ src/Ratatoskr.EfCore/Management/Endpoints/Inbox/DeleteInboxHandlerEndpoint.cs
@@ -2,6 +2,9 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Ratatoskr.EfCore.Internal;
 
 namespace Ratatoskr.EfCore.Management;
 
@@ -16,15 +19,45 @@ internal static class DeleteInboxHandlerEndpoint
         string contextName,
         Guid handlerStatusId,
         EfCoreManagementProviderLookup lookup,
+        IServiceScopeFactory scopeFactory,
         CancellationToken ct)
     {
         var provider = lookup.Find(contextName);
         if (provider is null || !provider.HasInbox) return TypedResults.NotFound();
 
-        var outcome = await provider.DeleteInboxHandlerStatusAsync(handlerStatusId, ct);
-        if (outcome == SingleDeleteOutcome.Success) return TypedResults.Ok();
-        if (outcome == SingleDeleteOutcome.NotFound) return TypedResults.NotFound();
-        if (outcome == SingleDeleteOutcome.NotPoisoned) return TypedResults.BadRequest("Handler status is not poisoned.");
-        return TypedResults.Conflict();
+        using var scope = scopeFactory.CreateScope();
+        var db = provider.GetDbContext(scope.ServiceProvider);
+
+        var entity = await db.Set<InboxHandlerStatusEntity>()
+            .SingleOrDefaultAsync(x => x.Id == handlerStatusId, ct);
+
+        if (entity is null) return TypedResults.NotFound();
+        if (!entity.IsPoisoned) return TypedResults.BadRequest("Handler status is not poisoned.");
+
+        var messageId = entity.MessageId;
+        db.Set<InboxHandlerStatusEntity>().Remove(entity);
+
+        // Atomically check whether this is the only remaining handler and delete the orphaned
+        // parent message in the same SaveChanges call to avoid a TOCTOU window.
+        var otherHandlerCount = await db.Set<InboxHandlerStatusEntity>()
+            .CountAsync(x => x.MessageId == messageId && x.Id != handlerStatusId, ct);
+
+        if (otherHandlerCount == 0)
+        {
+            var parent = await db.Set<InboxMessageEntity>()
+                .FindAsync([messageId], ct);
+            if (parent is not null)
+                db.Set<InboxMessageEntity>().Remove(parent);
+        }
+
+        try
+        {
+            await db.SaveChangesAsync(ct);
+            return TypedResults.Ok();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            return TypedResults.Conflict();
+        }
     }
 }
