+++ src/Ratatoskr.EfCore/Management/Endpoints/Inbox/GetInboxHandlerDetailEndpoint.cs
@@ -3,22 +3,11 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Ratatoskr.EfCore.Internal;
 
 namespace Ratatoskr.EfCore.Management;
-
-internal record InboxHandlerDetail(
-    Guid HandlerStatusId,
-    string MessageId,
-    string MessageType,
-    string HandlerKey,
-    DateTimeOffset ReceivedAt,
-    int ErrorCount,
-    int RequeuedCount,
-    string? LastError,
-    JsonElement Properties,
-    string? JsonPayload,
-    string PayloadBase64,
-    string DbContext);
 
 internal static class GetInboxHandlerDetailEndpoint
 {
@@ -31,12 +20,47 @@ internal static class GetInboxHandlerDetailEndpoint
         string contextName,
         Guid handlerStatusId,
         EfCoreManagementProviderLookup lookup,
+        IServiceScopeFactory scopeFactory,
         CancellationToken ct)
     {
         var provider = lookup.Find(contextName);
         if (provider is null || !provider.HasInbox) return TypedResults.NotFound();
 
-        var detail = await provider.GetPoisonedInboxDetailAsync(handlerStatusId, ct);
-        return detail is null ? TypedResults.NotFound() : TypedResults.Ok(detail);
+        using var scope = scopeFactory.CreateScope();
+        var db = provider.GetDbContext(scope.ServiceProvider);
+        db.ChangeTracker.QueryTrackingBehavior = QueryTrackingBehavior.NoTracking;
+
+        var result = await (
+            from hs in db.Set<InboxHandlerStatusEntity>()
+            join msg in db.Set<InboxMessageEntity>() on hs.MessageId equals msg.Id
+            where hs.Id == handlerStatusId && hs.IsPoisoned
+            select new { hs, msg }
+        ).FirstOrDefaultAsync(ct);
+
+        if (result is null) return TypedResults.NotFound();
+
+        var props = ManagementHelpers.SafeDeserializeToJsonElement(result.msg.SerializedProperties);
+        var msgType = ManagementHelpers.ExtractType(result.msg.SerializedProperties);
+        var (jsonPayload, base64) = ManagementHelpers.DecodeContent(result.msg.Content);
+
+        return TypedResults.Ok(new InboxHandlerDetail(
+            result.hs.Id, result.hs.MessageId, msgType, result.hs.HandlerKey, result.msg.ReceivedAt,
+            result.hs.ErrorCount, result.hs.RequeuedCount,
+            string.IsNullOrEmpty(result.hs.LastError) ? null : result.hs.LastError,
+            props, jsonPayload, base64, provider.DbContextName));
     }
+
+    internal record InboxHandlerDetail(
+        Guid HandlerStatusId,
+        string MessageId,
+        string MessageType,
+        string HandlerKey,
+        DateTimeOffset ReceivedAt,
+        int ErrorCount,
+        int RequeuedCount,
+        string? LastError,
+        JsonElement Properties,
+        string? JsonPayload,
+        string PayloadBase64,
+        string DbContext);
 }
