+++ src/Ratatoskr.EfCore/Management/Endpoints/Inbox/RequeueInboxHandlerEndpoint.cs
@@ -2,6 +2,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Ratatoskr.EfCore.Management;
 
@@ -16,12 +17,16 @@ internal static class RequeueInboxHandlerEndpoint
         string contextName,
         Guid handlerStatusId,
         EfCoreManagementProviderLookup lookup,
+        IServiceScopeFactory scopeFactory,
         CancellationToken ct)
     {
         var provider = lookup.Find(contextName);
         if (provider is null || !provider.HasInbox) return TypedResults.NotFound();
 
-        var outcome = await provider.RequeueInboxHandlerAsync(handlerStatusId, ct);
+        using var scope = scopeFactory.CreateScope();
+        var db = provider.GetDbContext(scope.ServiceProvider);
+
+        var outcome = await RequeueHelper.RequeueInboxHandlerAsync(db, handlerStatusId, ct);
         if (outcome == SingleRequeueOutcome.Success) return TypedResults.Ok();
         if (outcome == SingleRequeueOutcome.NotFound) return TypedResults.NotFound();
         if (outcome == SingleRequeueOutcome.NotPoisoned) return TypedResults.BadRequest("Handler status is not poisoned.");
