+++ src/Ratatoskr.EfCore/Management/Endpoints/Outbox/BulkDeleteOutboxEndpoint.cs
@@ -3,10 +3,11 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Ratatoskr.EfCore.Internal;
 
 namespace Ratatoskr.EfCore.Management;
-
-internal record BulkDeleteOutboxRequest(List<Guid>? Ids, bool? All);
 
 internal static class BulkDeleteOutboxEndpoint
 {
@@ -19,12 +20,32 @@ internal static class BulkDeleteOutboxEndpoint
         string contextName,
         [FromBody] BulkDeleteOutboxRequest req,
         EfCoreManagementProviderLookup lookup,
+        IServiceScopeFactory scopeFactory,
         CancellationToken ct)
     {
         var provider = lookup.Find(contextName);
         if (provider is null || !provider.HasOutbox) return TypedResults.NotFound();
 
-        await provider.BulkDeleteOutboxAsync(req.Ids, req.All is true, ct);
+        using var scope = scopeFactory.CreateScope();
+        var db = provider.GetDbContext(scope.ServiceProvider);
+
+        if (req.All is true)
+        {
+            await db.Set<OutboxMessageEntity>()
+                .Where(x => x.IsPoisoned)
+                .ExecuteDeleteAsync(ct);
+            return TypedResults.Ok();
+        }
+
+        if (req.Ids is not null and { Count: > 0 })
+        {
+            await db.Set<OutboxMessageEntity>()
+                .Where(x => req.Ids.Contains(x.Id) && x.IsPoisoned)
+                .ExecuteDeleteAsync(ct);
+        }
+
         return TypedResults.Ok();
     }
+
+    internal record BulkDeleteOutboxRequest(List<Guid>? Ids, bool? All);
 }
