+++ src/Ratatoskr.EfCore/Management/Endpoints/Outbox/DeleteOutboxEndpoint.cs
@@ -2,6 +2,9 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Routing;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.DependencyInjection;
+using Ratatoskr.EfCore.Internal;
 
 namespace Ratatoskr.EfCore.Management;
 
@@ -16,15 +19,30 @@ internal static class DeleteOutboxEndpoint
         string contextName,
         Guid id,
         EfCoreManagementProviderLookup lookup,
+        IServiceScopeFactory scopeFactory,
         CancellationToken ct)
     {
         var provider = lookup.Find(contextName);
         if (provider is null || !provider.HasOutbox) return TypedResults.NotFound();
 
-        var outcome = await provider.DeleteOutboxAsync(id, ct);
-        if (outcome == SingleDeleteOutcome.Success) return TypedResults.Ok();
-        if (outcome == SingleDeleteOutcome.NotFound) return TypedResults.NotFound();
-        if (outcome == SingleDeleteOutcome.NotPoisoned) return TypedResults.BadRequest("Message is not poisoned.");
-        return TypedResults.Conflict();
+        using var scope = scopeFactory.CreateScope();
+        var db = provider.GetDbContext(scope.ServiceProvider);
+
+        var entity = await db.Set<OutboxMessageEntity>()
+            .SingleOrDefaultAsync(x => x.Id == id, ct);
+
+        if (entity is null) return TypedResults.NotFound();
+        if (!entity.IsPoisoned) return TypedResults.BadRequest("Message is not poisoned.");
+
+        db.Set<OutboxMessageEntity>().Remove(entity);
+        try
+        {
+            await db.SaveChangesAsync(ct);
+            return TypedResults.Ok();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            return TypedResults.Conflict();
+        }
     }
 }
