+++ src/Ratatoskr.EfCore/Management/ManagementOutcomes.cs
@@ -3,5 +3,3 @@ namespace Ratatoskr.EfCore.Management;
 internal enum SingleRequeueOutcome { Success, NotFound, NotPoisoned, Conflict }
 
 internal enum SingleDeleteOutcome { Success, NotFound, NotPoisoned, Conflict }
-
-internal record RequeueMessageOutcome(bool Found, bool Conflict, IReadOnlyList<Guid> RequeuedIds);
